Add a tree class to handle commit messages
This tree class only allows insertion, never removal of objects, since
it is unlikely that we will remove a commit object.

It requires that each entry be unique, which should be fine now, since
I am planning to store SHA1 hashes from commit messages only.  Note
that the class does not validate this invariant in order to avoid
searching the entire tree on each insert.  In fact, insert should be
a O(1) operation.

Also, add a depth-first iterator to allow the tree to be traversed.