Add the commit time to the create method
Originally, I was going to have the commit factory create the time,
but that will not work, since I want the commit time to be when the user
started the operation, not when it completed.  The create method will
only be called when the operation is completed (since it needs the
output).  So we'll need to store the commit time when the opertion starts
and use it in the create call later.