Fix double equal comparison for some unequal numbers.
The C++ abs method takes either an int or a long.  Since we are
using an int64_t as the input, abs will not work correctly, returning
zero when the inputs are far from zero values apart.  Instead of using
abs, we'll check for both the positive and negative bounds.