Remove the osoasso_instance from the osoasso.c file.
It seems to be difficult to test pp::Var, as it is causing the test
executable to crash when its constructor is called.  So I have pulled the
real code for the instance out into separate files where it can be tested.
The connection code to JavaScript will not be tested, but it should remain
trivial.