Format matrices as HTML tables.
I had tried using MathJax with both Latex and MathML formating.
However, I was unable to get either to display in the same fixed-width
font as the rest of the GUI.  So I decided to fall back to allowing the
C++ code to format an HTML table that is styled in the HTML page to
use the correct font.