Change the timer to use gettimeofday() instead of std::clock()
The discussion here:

https://groups.google.com/group/native-client-discuss/browse_thread/thread/925ec6b67e4339a3/03282624c84423bc?lnk=gst&q=clock#03282624c84423bc

indicates that the std::clock() functionality was broken at one point.
It seems that it should be fixed in Chrome 16 and pepper 15, but it is
not.  I have followed the advice in the thread and switched to use
the gettimeofday() method instead of std::clock().  I have also changed
the unit test runner to use the timer class, so it will pick up the correct
implementation.