Change the multiply command to use row and column iterators.
Previously, the multiply command iterated the elements of each matrix.
It is now using the row and column iterators.  This seems to be signifigantly
faster, which is somewhat surprising.  The multiply stress test now takes less
than half the time it did before.  I believe that the results are still correct,
since the unit tests all pass.