Use SSE2 instructions to do vector multiplication
Change the matrix class to use an allocator that calls memalign to
align memory on 16 byte boundaries, as is required by SSE instructions.
Then use SSE2 intrinsics to perform the multiplication two doubles at a time.

Initially, it does not seem that this is any faster than row and column based
multiplication, although I'll need to investigate further.