Add clobbered registers to the inline assembly
Previously, I did not know that I need to inform gcc about the registers
I will be using in the assembly code.  So it was attempting to use some of the
registers, and I was stepping on them.  Also, I have added some tests for larger
matrices to ensure that I am getting correct results.